pyproject.toml
[tool.poetry]
name = "homework2"
version = "0.1.0"
description = ""
authors = ["stanislav <stasfed16@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.8.0"
mypy = "^1.11.2"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
